Refactor tool handling functions to return structured results
- Updated `handleValidate`, `handleHealthCheck`, and `handleDiscover` functions to return an object containing both the output text and an error status instead of just a string.
- Introduced a new `ToolExecutionResult` interface to standardize the return type.
- Modified input schemas in `tools.ts` to include new validation rules for tokens and client profiles.
- Added comprehensive tests for the updated handler functions to ensure correct behavior with mock CLI outputs and error handling.

+++ Mcp.Benchmark.Infrastructure/Utilities/PerformanceMetricsCalculator.cs
@@ -12,7 +12,7 @@ public static class PerformanceMetricsCalculator
     /// </summary>
     /// <param name="times">Response times (will be sorted internally).</param>
     /// <param name="percentile">Percentile to calculate (0.0 to 1.0, e.g., 0.95 for P95).</param>
-    public static double GetPercentile(IReadOnlyList<long> times, double percentile)
+    public static double GetPercentile(IReadOnlyList<double> times, double percentile)
     {
         if (times.Count == 0) return 0;
         var sorted = times.OrderBy(x => x).ToList();
